requirements.txt
folium==0.18.0
haversine==2.8.1
matplotlib==3.8.4
numpy>=1.24.0
pandas==2.2.3
Pillow==10.0.0
plotly==5.22.0
streamlit==1.37.1
streamlit_folium==0.13.0